requirements.txt
appnope==0.1.0
Babel==2.5.1
boto3==1.5.6
decorator==4.0.11
dj-database-url==0.4.2
Django==1.11
django-js-asset==0.1.1
django-localflavor==1.4.1
django-phonenumber-field==1.3.0
django-registration==2.2
djangorestframework==3.7.7
django-storages==1.6.5
django-wysiwyg-redactor==0.5.1
gunicorn==19.7.1
ipython==6.0.0
ipython-genutils==0.2.0
jedi==0.10.2
olefile==0.44
passlib==1.7.1
pexpect==4.2.1
phonenumberslite==8.8.7
pickleshare==0.7.4
Pillow==4.1.0
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
Pygments==2.2.0
pytz==2017.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
